Connect School/ConnectWeb/requirements.txt
Django==2.0.4
django-filter==1.1.0
django-model-utils==3.1.1
django-polymorphic==2.0.2
django-rest-polymorphic==0.1.5
djangorestframework==3.8.2
gunicorn==19.7.1
Markdown==2.6.11
psycopg2==2.7.4
pytz==2018.4
six==1.11.0
whitenoise==3.3.1
